+++ src/UserBundle/Form/ProfileType.php
@@ -1,17 +1,20 @@
 <?php
 
-namespace UserBundle\Form\Type;
+namespace UserBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 
-class RegistrationType extends AbstractType
+class ProfileType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder->add('name', null, array(
+        $builder
+            ->add('name', null, array(
             'label' => 'Votre nom'))
+            ->remove('current_password') // suppression champ mdp actuel
+
             ->add('newsletter', CheckboxType::class, array(
                 'label'    => 'Souhaitez-vous recevoir la newsletter ?',
                 'required' => false,
@@ -20,7 +23,7 @@ class RegistrationType extends AbstractType
 
     public function getParent()
     {
-        return 'FOS\UserBundle\Form\Type\RegistrationFormType';
+        return 'FOS\UserBundle\Form\Type\ProfileFormType';
 
         // Or for Symfony < 2.8
         // return 'fos_user_registration';
@@ -28,7 +31,7 @@ class RegistrationType extends AbstractType
 
     public function getBlockPrefix()
     {
-        return 'app_user_registration';
+        return 'app_user_profile';
     }
 
     // For Symfony 2.x
