+++ src/UserBundle/Form/RegistrationType.php
@@ -1,20 +1,17 @@
 <?php
 
-namespace UserBundle\Form\Type;
+namespace UserBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\FormBuilderInterface;
 
-class ProfileType extends AbstractType
+class RegistrationType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder
-            ->add('name', null, array(
+        $builder->add('name', null, array(
             'label' => 'Votre nom'))
-            ->remove('current_password') // suppression champ mdp actuel
-
             ->add('newsletter', CheckboxType::class, array(
                 'label'    => 'Souhaitez-vous recevoir la newsletter ?',
                 'required' => false,
@@ -23,15 +20,12 @@ class ProfileType extends AbstractType
 
     public function getParent()
     {
-        return 'FOS\UserBundle\Form\Type\ProfileFormType';
-
-        // Or for Symfony < 2.8
-        // return 'fos_user_registration';
+        return 'FOS\UserBundle\Form\Type\RegistrationFormType';
     }
 
     public function getBlockPrefix()
     {
-        return 'app_user_profile';
+        return 'app_user_registration';
     }
 
     // For Symfony 2.x
