+++ src/UserBundle/Form/SearchType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace UserBundle\Form\Type;
+namespace UserBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
